Fix connection-liveness testing.
I disabled connection-liveness testing previously because I didn't want
to deal with plumbing the underlying connection around as well as the
TLS connection built on top of it--the underlying connection is
necessary to make syscalls directly, which is necessary for liveness
checking, and the crypto/tls package does not provide a way to retrieve
the underlying connection for an arbitrary TLS connection.

Note that if the underlying net.Conn is not a syscall.Conn (i.e. it's
not implemented as something that is interacted with via syscalls,
more-or-less), connection liveness checking is still disabled. I'm not
sure there is a good way to do liveness checking in this case.