Initial full implementation of hrelayd.
The hrelayd binary was added long ago, but it was just a few lines that
sketched how to create an hrelay server. Now it's a full-featured
implementation. The server gets its certificate from Let's Encrypt (and
so clients are expected to use standard Internet PKI CAs to evaluate
trust of the server). Client certificates are expected to be signed by a
user-configured set of CAs, whose certificates must be passed to the
server.